python/flask/requirements.txt
# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# REST API with OpenAPI/Swagger
flask-restx==1.3.0

# GraphQL
graphene==3.3
Flask-GraphQL==2.0.1

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Database
alembic==1.13.1
SQLAlchemy==2.0.25

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Observability & Monitoring
prometheus-flask-exporter==0.23.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Logging
python-json-logger==2.0.7

# Configuration
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Security
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# CLI
click==8.1.7
requests==2.31.0
rich==13.7.0
prompt-toolkit==3.0.43

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.12.1
flake8==7.0.0
